+++ backend/.env
@@ -1,2 +1,2 @@
-# mongoUri=<mongoDB URI>
-PORT=<PORT NUMBER>
+mongoUri=mongodb+srv://picsmash7:MrP6Jb1q5MMlvOD3@merncluster.semrwea.mongodb.net/?retryWrites=true&w=majority&appName=MernCluster
+PORT=5001
