+++ backend/models/User.js
@@ -1,0 +1,16 @@
+const mongoose = require("mongoose");
+
+const UserSchema = new mongoose.Schema({
+  name: {
+    type: String,
+    required: true,
+  },
+  email: {
+    type: String,
+    required: true,
+    unique: true,
+  },
+  // 다른 필드들...
+});
+
+module.exports = mongoose.model("User", UserSchema);
