+++ backend/server.js
@@ -5,9 +5,9 @@ const connectDB = require("./config/db");
 const dotenv = require("dotenv");
 // const axios = require("axios");
 const cors = require("cors");
-const PORT = process.env.PORT || 5000;
+const PORT = process.env.PORT || 5001;
+console.log("PORT: " + PORT);
 
-// app.listen(port, () => console.log("running"));
 // express 는 함수이므로, 반환값을 변수에 저장한다.
 const app = express();
 connectDB();
@@ -20,10 +20,6 @@ app.use(
     extended: false,
   })
 );
-
-app.use("/api/user", require("./routes/api/user"));
-app.use("/api/auth", require("./routes/api/auth"));
-
 // node.js에서 json 파일을 사용할때 필요(필수는 아닌듯)
 app.use(
   express.json({
@@ -33,36 +29,8 @@ app.use(
 
 app.use(cors()); // 모든 도메인에서의 요청을 허용
 // 5001 포트로 서버 오픈
-app.listen(5001, () => {
-  console.log("start! express server on port 5001");
+app.listen(PORT, () => {
+  console.log(`start! express server on port ${PORT}`);
 });
 
-app.use("/api/user", require("./routes/api/user"));
-app.use("/api/auth", require("./routes/api/auth"));
-
-app.get("/", function (req, res) {
-  res.send({ message: "Hello world!" });
-});
-
-app.get("/test", function (req, res) {
-  res.send(
-    `
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>main.html</title>
-</head>
-<body>
-    <h1>main page</h1>
-
-    <p>Lorem, ipsum dolor sit amet consectetur adipisicing elit. 
-        Atque, ullam quos! Iste quae, molestiae vitae tenetur, 
-        in a aperiam voluptatibus dicta qui doloribus libero suscipit optio delectus voluptas voluptatem impedit!</p>
-</body>
-</html>
-`
-  );
-});
+app.use("/api/users", require("./routes/api/users")); // 경로에 맞게 조정
